# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Healthcare Recruiting ATS/CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates/{id}/tags:
    post:
      operationId: addCandidateTag
      tags: [tags]
      summary: Add a tag to a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Tag added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /candidates/{id}/tags/{tagId}:
    delete:
      operationId: removeCandidateTag
      tags: [tags]
      summary: Remove a tag from a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: tagId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # PIPELINE
...
components:
  schemas:
...
    Tag:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: ["string", "null"]
    TagInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        color:
          type: string

    # PIPELINE
...