# lib/api-zod/src/generated/types/tagInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Healthcare Recruiting ATS/CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface TagInput {
  name: string;
  color?: string;
}